app/requirements.txt
alembic==1.12.0
fastapi==0.110.0
pydantic==2.6.4
pydantic_settings==2.2.1
PyJWT==2.8.0
PyJWT==2.8.0
python-dotenv==1.0.1
Requests==2.31.0
SQLAlchemy==2.0.23
uvicorn==0.28.0
asyncpg